Review Endpoints + fixes + Implement Authenticated User Access Endpoints

+++ part3/app/api/v1/reviews.py
@@ -1,3 +1,4 @@
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from flask_restx import Namespace, Resource, fields
 from app.services import HBnBFacade
 
@@ -14,6 +15,7 @@ review_model = api.model('Review', {
 
 @api.route('/')
 class ReviewList(Resource):
+    @jwt_required() # Requires a JWT token to access this route
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
@@ -21,6 +23,25 @@ class ReviewList(Resource):
         """Register a new review"""
         if not api.payload:
             return {'error': 'Request payload is missing or invalid'}, 400
+        
+        current_user = get_jwt_identity()  # Recover logged-in user ID
+        review_data = api.payload
+        place_id = review_data.get("place_id")
+
+        # Check that the location exists
+        place = facade.get_place(place_id)
+        if not place:
+            return {'error': "Place not found"}, 404
+        
+        # Prevent an owner from rating his own location
+        if place["owner_id"] == current_user:
+            return {'error': "You cannot review your own place"}, 403
+        
+        # Check if the user has already left a review for this location
+        existing_review = facade.get_review_by_user_and_place(current_user, place_id)
+        if existing_review:
+            return {'error': "You have already reviewed this place"}, 403
+        
         try:
             review_data = api.payload
             new_review = facade.create_review(review_data)
@@ -49,6 +70,7 @@ class ReviewResource(Resource):
         except ValueError as e:
             return {'error': str(e)}, 404
 
+    @jwt_required() # Requires a JWT token to access this route
     @api.expect(review_model)
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
@@ -57,6 +79,18 @@ class ReviewResource(Resource):
         """Update a review's information"""
         if not api.payload:
             return {'error': "Request payload is missing or invalid"}, 400
+        
+        current_user = get_jwt_identity() # Recover logged-in user ID
+
+        # Check that review exists
+        review = facade.get_review(review_id)
+        if not review:
+            return {'error': "Review not found"}, 404
+
+        # Check if the user is the author of the review
+        if review["user_id"] != current_user:
+            return {'error': "You are not authorized to update this review"}, 403
+
         try:
             updated_data = api.payload
             updated_review = facade.update_review(review_id, updated_data)
@@ -64,10 +98,23 @@ class ReviewResource(Resource):
         except ValueError as e:
             return {'error': str(e)}, 400
 
+    @jwt_required() # Requires a JWT token to access this route
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @api.response(403, 'Unauthorized action')
     def delete(self, review_id):
         """Delete a review"""
+        current_user = get_jwt_identity()  # Recover logged-in user ID
+
+        # Check that review exists
+        review = facade.get_review(review_id)
+        if not review:
+            return {'error': "Review not found"}, 404
+
+        # Check if the user is the author of the review
+        if review["user_id"] != current_user:
+            return {'error': "You are not authorized to delete this review"}, 403
+                    
         try:
             facade.delete_review(review_id)
             return {'message': "Review successfully deleted"}, 200
